backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
numpy>=1.26.0
redis>=5.0.1
qdrant-client>=1.6.4
openai>=1.3.5
loguru>=0.7.2
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
pandas
ftfy
git+https://github.com/openai/CLIP.git
python-multipart==0.0.20
torch>=2.0.0
Pillow>=10.0.0
instaloader>=4.10.0
aiofiles>=23.0.0
PyJWT>=2.8.0
jinja2>=3.0.0
SQLAlchemy==2.0.41
psycopg2==2.9.10
websockets>=11.0.3
pytz>=2023.3